fix driver/vus filter and create store for comparison and results view store overlap
