+++ src/pages/resultsView/ResultsViewPageStore.ts
@@ -303,6 +303,7 @@ import { getAlterationData } from 'shared/components/oncoprint/OncoprintUtils';
 import { PageUserSession } from 'shared/userSession/PageUserSession';
 import { PageType } from 'shared/userSession/PageType';
 import { ClinicalTrackConfig } from 'shared/components/oncoprint/Oncoprint';
+import AnalysisStore from 'shared/lib/comparison/AnalysisStore';
 
 type Optional<T> =
     | { isApplicable: true; value: T }
@@ -544,21 +545,20 @@ interface IResultsViewExclusionSettings {
 /* fields and methods in the class below are ordered based on roughly
 /* chronological setup concerns, rather than on encapsulation and public API */
 /* tslint:disable: member-ordering */
-export class ResultsViewPageStore
+export class ResultsViewPageStore extends AnalysisStore
     implements
         IAnnotationFilterSettings,
         IResultsViewExclusionSettings,
         ISettingsMenuButtonVisible {
-    @observable driverAnnotationSettings: DriverAnnotationSettings;
-
     private reactionDisposers: IReactionDisposer[] = [];
 
     public pageUserSession: PageUserSession<ResultPageSettings>;
 
     constructor(
-        private appStore: AppStore,
+        protected appStore: AppStore,
         public urlWrapper: ResultsViewURLWrapper
     ) {
+        super(appStore);
         makeObservable(this);
         this.getURL();
 
@@ -3797,47 +3797,6 @@ export class ResultsViewPageStore
         }`;
     }
 
-    readonly oncoKbCancerGenes = remoteData(
-        {
-            invoke: () => {
-                if (getServerConfig().show_oncokb) {
-                    return fetchOncoKbCancerGenes();
-                } else {
-                    return Promise.resolve([]);
-                }
-            },
-        },
-        []
-    );
-
-    readonly oncoKbAnnotatedGenes = remoteData(
-        {
-            await: () => [this.oncoKbCancerGenes],
-            invoke: () => {
-                if (getServerConfig().show_oncokb) {
-                    return Promise.resolve(
-                        _.reduce(
-                            this.oncoKbCancerGenes.result,
-                            (
-                                map: { [entrezGeneId: number]: boolean },
-                                next: CancerGene
-                            ) => {
-                                if (next.oncokbAnnotated) {
-                                    map[next.entrezGeneId] = true;
-                                }
-                                return map;
-                            },
-                            {}
-                        )
-                    );
-                } else {
-                    return Promise.resolve({});
-                }
-            },
-        },
-        {}
-    );
-
     readonly clinicalDataForSamples = remoteData<ClinicalData[]>(
         {
             await: () => [this.studies, this.samples],
@@ -4289,13 +4248,6 @@ export class ResultsViewPageStore
         return getGenomeBuildFromStudies(this.studies.result);
     }
 
-    @computed get referenceGenomeBuild() {
-        if (!this.studies.isComplete) {
-            throw new Error('Failed to get studies');
-        }
-        return getGenomeNexusUrl(this.studies.result);
-    }
-
     @autobind
     generateGenomeNexusHgvsgUrl(hgvsg: string) {
         return getGenomeNexusHgvsgUrl(hgvsg, this.referenceGenomeBuild);
@@ -4310,10 +4262,6 @@ export class ResultsViewPageStore
 
     @computed get genomeNexusClient() {
         return new GenomeNexusAPI(this.referenceGenomeBuild);
-    }
-
-    @computed get genomeNexusInternalClient() {
-        return new GenomeNexusAPIInternal(this.referenceGenomeBuild);
     }
 
     //this is only required to show study name and description on the results page
@@ -4826,14 +4774,6 @@ export class ResultsViewPageStore
         },
     });
 
-    readonly entrezGeneIdToGene = remoteData<{ [entrezGeneId: number]: Gene }>({
-        await: () => [this.genes],
-        invoke: () =>
-            Promise.resolve(
-                _.keyBy(this.genes.result!, gene => gene.entrezGeneId)
-            ),
-    });
-
     readonly genesetLinkMap = remoteData<{ [genesetId: string]: string }>({
         invoke: async () => {
             if (this.genesetIds && this.genesetIds.length) {
@@ -5043,23 +4983,6 @@ export class ResultsViewPageStore
         }
     );
 
-    readonly _filteredAndAnnotatedMutationsReport = remoteData({
-        await: () => [
-            this.mutations,
-            this.getMutationPutativeDriverInfo,
-            this.entrezGeneIdToGene,
-        ],
-        invoke: () => {
-            return Promise.resolve(
-                filterAndAnnotateMutations(
-                    this.mutations.result!,
-                    this.getMutationPutativeDriverInfo.result!,
-                    this.entrezGeneIdToGene.result!
-                )
-            );
-        },
-    });
-
     readonly _filteredAndAnnotatedStructuralVariantsReport = remoteData({
         await: () => [
             this.structuralVariants,
@@ -5070,27 +4993,6 @@ export class ResultsViewPageStore
                 filterAndAnnotateStructuralVariants(
                     this.structuralVariants.result!,
                     this.getStructuralVariantPutativeDriverInfo.result!
-                )
-            );
-        },
-    });
-
-    readonly filteredAndAnnotatedMutations = remoteData<AnnotatedMutation[]>({
-        await: () => [
-            this._filteredAndAnnotatedMutationsReport,
-            this.filteredSampleKeyToSample,
-        ],
-        invoke: () => {
-            const filteredMutations = compileMutations(
-                this._filteredAndAnnotatedMutationsReport.result!,
-                !this.driverAnnotationSettings.includeVUS,
-                !this.includeGermlineMutations
-            );
-            const filteredSampleKeyToSample = this.filteredSampleKeyToSample
-                .result!;
-            return Promise.resolve(
-                filteredMutations.filter(
-                    m => m.uniqueSampleKey in filteredSampleKeyToSample
                 )
             );
         },
@@ -5221,78 +5123,6 @@ export class ResultsViewPageStore
         },
     }));
 
-    readonly getMutationPutativeDriverInfo = remoteData({
-        await: () => {
-            const toAwait = [];
-            if (this.driverAnnotationSettings.oncoKb) {
-                toAwait.push(this.oncoKbMutationAnnotationForOncoprint);
-            }
-            if (this.driverAnnotationSettings.hotspots) {
-                toAwait.push(this.isHotspotForOncoprint);
-            }
-            if (this.driverAnnotationSettings.cbioportalCount) {
-                toAwait.push(this.getCBioportalCount);
-            }
-            if (this.driverAnnotationSettings.cosmicCount) {
-                toAwait.push(this.getCosmicCount);
-            }
-            return toAwait;
-        },
-        invoke: () => {
-            return Promise.resolve((mutation: Mutation): {
-                oncoKb: string;
-                hotspots: boolean;
-                cbioportalCount: boolean;
-                cosmicCount: boolean;
-                customDriverBinary: boolean;
-                customDriverTier?: string;
-            } => {
-                const getOncoKbMutationAnnotationForOncoprint = this
-                    .oncoKbMutationAnnotationForOncoprint.result!;
-                const oncoKbDatum:
-                    | IndicatorQueryResp
-                    | undefined
-                    | null
-                    | false =
-                    this.driverAnnotationSettings.oncoKb &&
-                    getOncoKbMutationAnnotationForOncoprint &&
-                    !(
-                        getOncoKbMutationAnnotationForOncoprint instanceof Error
-                    ) &&
-                    getOncoKbMutationAnnotationForOncoprint(mutation);
-
-                const isHotspotDriver =
-                    this.driverAnnotationSettings.hotspots &&
-                    !(this.isHotspotForOncoprint.result instanceof Error) &&
-                    this.isHotspotForOncoprint.result!(mutation);
-                const cbioportalCountExceeded =
-                    this.driverAnnotationSettings.cbioportalCount &&
-                    this.getCBioportalCount.isComplete &&
-                    this.getCBioportalCount.result!(mutation) >=
-                        this.driverAnnotationSettings.cbioportalCountThreshold;
-                const cosmicCountExceeded =
-                    this.driverAnnotationSettings.cosmicCount &&
-                    this.getCosmicCount.isComplete &&
-                    this.getCosmicCount.result!(mutation) >=
-                        this.driverAnnotationSettings.cosmicCountThreshold;
-
-                // Note: custom driver annotations are part of the incoming datum
-                return evaluateMutationPutativeDriverInfo(
-                    mutation,
-                    oncoKbDatum,
-                    this.driverAnnotationSettings.hotspots,
-                    isHotspotDriver,
-                    this.driverAnnotationSettings.cbioportalCount,
-                    cbioportalCountExceeded,
-                    this.driverAnnotationSettings.cosmicCount,
-                    cosmicCountExceeded,
-                    this.driverAnnotationSettings.customBinary,
-                    this.driverAnnotationSettings.driverTiers
-                );
-            });
-        },
-    });
-
     readonly getStructuralVariantPutativeDriverInfo = remoteData({
         await: () => {
             const toAwait = [];
@@ -5417,28 +5247,6 @@ export class ResultsViewPageStore
         undefined
     );
 
-    // Hotspots
-    readonly hotspotData = remoteData({
-        await: () => [this.mutations],
-        invoke: () => {
-            return fetchHotspotsData(
-                this.mutations,
-                undefined,
-                this.genomeNexusInternalClient
-            );
-        },
-    });
-
-    readonly indexedHotspotData = remoteData<IHotspotIndex | undefined>({
-        await: () => [this.hotspotData],
-        invoke: () => Promise.resolve(indexHotspotsData(this.hotspotData)),
-    });
-
-    public readonly isHotspotForOncoprint = remoteData<
-        ((m: Mutation) => boolean) | Error
-    >({
-        invoke: () => makeIsHotspotForOncoprint(this.indexedHotspotData),
-    });
     //OncoKb
     readonly uniqueSampleKeyToTumorType = remoteData<{
         [uniqueSampleKey: string]: string;
@@ -5458,23 +5266,6 @@ export class ResultsViewPageStore
             );
         },
     });
-
-    //we need seperate oncokb data because oncoprint requires onkb queries across cancertype
-    //mutations tab the opposite
-    readonly oncoKbDataForOncoprint = remoteData<IOncoKbData | Error>(
-        {
-            await: () => [this.mutations, this.oncoKbAnnotatedGenes],
-            invoke: async () =>
-                fetchOncoKbDataForOncoprint(
-                    this.oncoKbAnnotatedGenes,
-                    this.mutations
-                ),
-            onError: (err: Error) => {
-                // fail silently, leave the error handling responsibility to the data consumer
-            },
-        },
-        ONCOKB_DEFAULT
-    );
 
     readonly structuralVariantOncoKbDataForOncoprint = remoteData<
         IOncoKbData | Error
@@ -5539,16 +5330,6 @@ export class ResultsViewPageStore
         );
     }
 
-    readonly oncoKbMutationAnnotationForOncoprint = remoteData<
-        Error | ((mutation: Mutation) => IndicatorQueryResp | undefined)
-    >({
-        await: () => [this.oncoKbDataForOncoprint],
-        invoke: () =>
-            makeGetOncoKbMutationAnnotationForOncoprint(
-                this.oncoKbDataForOncoprint
-            ),
-    });
-
     readonly oncoKbStructuralVariantAnnotationForOncoprint = remoteData<
         | Error
         | ((
@@ -5592,109 +5373,6 @@ export class ResultsViewPageStore
                 this.cnaOncoKbDataForOncoprint,
                 this.driverAnnotationSettings.oncoKb
             ),
-    });
-
-    readonly cbioportalMutationCountData = remoteData<{
-        [mutationCountByPositionKey: string]: number;
-    }>({
-        await: () => [this.mutations],
-        invoke: async () => {
-            const mutationPositionIdentifiers = _.values(
-                countMutations(this.mutations.result!)
-            );
-
-            if (mutationPositionIdentifiers.length > 0) {
-                const data = await internalClient.fetchMutationCountsByPositionUsingPOST(
-                    {
-                        mutationPositionIdentifiers,
-                    }
-                );
-                return _.mapValues(
-                    _.groupBy(data, mutationCountByPositionKey),
-                    (counts: MutationCountByPosition[]) =>
-                        _.sumBy(counts, c => c.count)
-                );
-            } else {
-                return {};
-            }
-        },
-    });
-
-    readonly getCBioportalCount: MobxPromise<
-        (mutation: Mutation) => number
-    > = remoteData({
-        await: () => [this.cbioportalMutationCountData],
-        invoke: () => {
-            return Promise.resolve((mutation: Mutation): number => {
-                const key = mutationCountByPositionKey(mutation);
-                return this.cbioportalMutationCountData.result![key] || -1;
-            });
-        },
-    });
-    //COSMIC count
-    readonly cosmicCountsByKeywordAndStart = remoteData<ComplexKeyCounter>({
-        await: () => [this.mutations],
-        invoke: async () => {
-            const keywords = _.uniq(
-                this.mutations
-                    .result!.filter((m: Mutation) => {
-                        // keyword is what we use to query COSMIC count with, so we need
-                        //  the unique list of mutation keywords to query. If a mutation has
-                        //  no keyword, it cannot be queried for.
-                        return !!m.keyword;
-                    })
-                    .map((m: Mutation) => m.keyword)
-            );
-
-            if (keywords.length > 0) {
-                const data = await internalClient.fetchCosmicCountsUsingPOST({
-                    keywords,
-                });
-                const map = new ComplexKeyCounter();
-                for (const d of data) {
-                    const position = getProteinPositionFromProteinChange(
-                        d.proteinChange
-                    );
-                    if (position) {
-                        map.add(
-                            {
-                                keyword: d.keyword,
-                                start: position.start,
-                            },
-                            d.count
-                        );
-                    }
-                }
-                return map;
-            } else {
-                return new ComplexKeyCounter();
-            }
-        },
-    });
-
-    readonly getCosmicCount: MobxPromise<
-        (mutation: Mutation) => number
-    > = remoteData({
-        await: () => [this.cosmicCountsByKeywordAndStart],
-        invoke: () => {
-            return Promise.resolve((mutation: Mutation): number => {
-                const targetPosObj = getProteinPositionFromProteinChange(
-                    mutation.proteinChange
-                );
-                if (targetPosObj) {
-                    const keyword = mutation.keyword;
-                    const cosmicCount = this.cosmicCountsByKeywordAndStart.result!.get(
-                        {
-                            keyword,
-                            start: targetPosObj.start,
-                        }
-                    );
-                    return cosmicCount;
-                } else {
-                    return -1;
-                }
-            });
-        },
     });
 
     readonly molecularProfileIdToProfiledFilteredSamples = remoteData({
@@ -5873,7 +5551,6 @@ export class ResultsViewPageStore
             ),
     });
 
-    readonly geneCache = new GeneCache();
     readonly genesetCache = new GenesetCache();
 
     private _numericGeneMolecularDataCache = new MobxPromiseCache<
