+++ src/shared/lib/comparison/ComparisonStore.ts
@@ -104,23 +104,20 @@ import {
     SessionGroupData,
 } from 'shared/api/session-service/sessionServiceModels';
 import { AlterationEnrichmentRow } from 'shared/model/AlterationEnrichmentRow';
+import AnalysisStore from './AnalysisStore';
 
 export enum OverlapStrategy {
     INCLUDE = 'Include',
     EXCLUDE = 'Exclude',
 }
 
-export default abstract class ComparisonStore
+export default abstract class ComparisonStore extends AnalysisStore
     implements IAnnotationFilterSettings {
     private tabHasBeenShown = observable.map<GroupComparisonTab, boolean>();
 
     private tabHasBeenShownReactionDisposer: IReactionDisposer;
     @observable public newSessionPending = false;
 
-    @observable
-    driverAnnotationSettings: DriverAnnotationSettings = buildDriverAnnotationSettings(
-        () => false
-    );
     @observable includeGermlineMutations = true;
     @observable includeSomaticMutations = true;
     @observable includeUnknownStatusMutations = true;
@@ -132,7 +129,12 @@ export default abstract class ComparisonStore
         protected urlWrapper: IComparisonURLWrapper,
         protected resultsViewStore?: ResultsViewPageStore
     ) {
+        super(appStore);
         makeObservable(this);
+
+        this.driverAnnotationSettings = buildDriverAnnotationSettings(
+            () => false
+        );
 
         (window as any).compStore = this;
 
@@ -273,7 +275,6 @@ export default abstract class ComparisonStore
     abstract get overlapStrategy(): OverlapStrategy;
     abstract get usePatientLevelEnrichments(): boolean;
     abstract get samples(): MobxPromise<Sample[]>;
-    abstract get studies(): MobxPromise<CancerStudy[]>;
     // < / >
 
     public get isLoggedIn() {
